+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated `ToUpperFirstLetter` and `ToLowerFirstLetter` and renamed as `ToUpperFirst` and `ToLowerFirst` (`Strings` class).
 - Renamed `Strings::IsWhiteSpace` as `IsWhitespace`.
 - Added `MaybeExtensions::ToJust` method.
+- Updated `Maybe::Equals(object)` implementation.
 
 ## [1.0.5] - 2021-11-28